+++ model/stylegan/op/__init__.py
@@ -1,2 +1,2 @@
 from .fused_act import FusedLeakyReLU, fused_leaky_relu
-from .upfirdn2d import upfirdn2d
+from .upfirdn2d_pkg import upfirdn2d
