+++ model/stylegan/op_cpu/upfirdn2d_pkg.py
