+++ model/stylegan/stylegan_model.py
@@ -8,7 +8,15 @@ from torch import nn
 from torch.nn import functional as F
 from torch.autograd import Function
 
-from model.stylegan.op_cpu import FusedLeakyReLU, fused_leaky_relu, upfirdn2d, conv2d_gradfix
+import os
+if 'USING_TORCHSERVE' in os.environ:
+    from fused_act import FusedLeakyReLU, fused_leaky_relu
+    from upfirdn2d_pkg import upfirdn2d
+    from conv2d_gradfix import conv2d
+    from conv2d_gradfix import conv_transpose2d
+else:
+    from model.stylegan.op_cpu import FusedLeakyReLU, fused_leaky_relu, upfirdn2d, conv2d_gradfix
+    from model.stylegan.op_cpu.conv2d_gradfix import conv2d, conv_transpose2d
 
 class PixelNorm(nn.Module):
     def __init__(self):
@@ -112,7 +120,7 @@ class EqualConv2d(nn.Module):
             self.bias = None
 
     def forward(self, input):
-        out = conv2d_gradfix.conv2d(
+        out = conv2d(
             input,
             self.weight * self.scale,
             bias=self.bias,
@@ -239,17 +247,17 @@ class ModulatedConv2d(nn.Module):
 
             if self.upsample:
                 weight = weight.transpose(0, 1)
-                out = conv2d_gradfix.conv_transpose2d(
+                out = conv_transpose2d(
                     input, weight, padding=0, stride=2
                 )
                 out = self.blur(out)
 
             elif self.downsample:
                 input = self.blur(input)
-                out = conv2d_gradfix.conv2d(input, weight, padding=0, stride=2)
+                out = conv2d(input, weight, padding=0, stride=2)
 
             else:
-                out = conv2d_gradfix.conv2d(input, weight, padding=self.padding)
+                out = conv2d(input, weight, padding=self.padding)
 
             if self.demodulate:
                 out = out * dcoefs.view(batch, -1, 1, 1)
@@ -278,7 +286,7 @@ class ModulatedConv2d(nn.Module):
             weight = weight.transpose(1, 2).reshape(
                 batch * in_channel, self.out_channel, self.kernel_size, self.kernel_size
             )
-            out = conv2d_gradfix.conv_transpose2d(
+            out = conv_transpose2d(
                 input, weight, padding=0, stride=2, groups=batch
             )
             _, _, height, width = out.shape
@@ -289,7 +297,7 @@ class ModulatedConv2d(nn.Module):
             input = self.blur(input)
             _, _, height, width = input.shape
             input = input.view(1, batch * in_channel, height, width)
-            out = conv2d_gradfix.conv2d(
+            out = conv2d(
                 input, weight, padding=0, stride=2, groups=batch
             )
             _, _, height, width = out.shape
@@ -297,7 +305,7 @@ class ModulatedConv2d(nn.Module):
 
         else:
             input = input.view(1, batch * in_channel, height, width)
-            out = conv2d_gradfix.conv2d(
+            out = conv2d(
                 input, weight, padding=self.padding, groups=batch
             )
             _, _, height, width = out.shape
